backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-cleanup>=8.0
python-decouple>=3.8
psycopg2-binary>=2.9
gunicorn>=21.2
reportlab>=4.0.0
openpyxl>=3.1.0
pyotp>=2.9.0
channels>=4.0
channels-redis>=4.0
